updated ERC20 OGToken

+++ src/ManualToken.sol
@@ -2,7 +2,7 @@
 pragma solidity ^0.8.18;
 
 contract ManualToken {
-    mapping(address => uint) private s_balances;
+    mapping(address => uint256) private s_balances;
 
     function name() public pure returns (string memory) {
         return "Glitch Token";
